Fixed merge conflict errors, updated upstream diff comments.

+++ ReactCommon/cxxreact/SampleCxxModule.cpp
@@ -28,9 +28,9 @@ std::string Sample::concat(const std::string& a, const std::string& b) {
   return a + b;
 }
 
-std::string Sample::repeat(int64_t count, const std::string& str) {
+std::string Sample::repeat(int count, const std::string& str) {
   std::string ret;
-  for (int64_t i = 0; i < count; i++) {
+  for (int i = 0; i < count; i++) {
     ret += str;
   }
